+++ R/do_CorrelationPlot.R
@@ -7,7 +7,8 @@
 #' @export
 #'
 #' @example /man/examples/examples_do_CorrelationPlot.R
-do_CorrelationPlot <- function(sample,
+do_CorrelationPlot <- function(sample = NULL,
+                               input_gene_list = NULL,
                                mode = "hvg",
                                assay = NULL,
                                group.by = NULL,
@@ -31,10 +32,14 @@ do_CorrelationPlot <- function(sample,
                                plot.subtitle = NULL,
                                plot.caption = NULL,
                                diverging.palette = "RdBu",
+                               use_viridis = FALSE,
+                               viridis.palette = "G",
+                               viridis_direction = -1,
+                               sequential.palette = "YlGnBu",
+                               sequential_direction = 1,
                                rotate_x_axis_labels = 45){
 
-  # Check if the sample provided is a Seurat object.
-  check_Seurat(sample = sample)
+ 
   # Check logical parameters.
   logical_list <- list("enforce_symmetry" = enforce_symmetry)
   check_type(parameters = logical_list, required_type = "logical", test_function = is.logical)
@@ -47,7 +52,9 @@ do_CorrelationPlot <- function(sample,
                        "legend.tickwidth" = legend.tickwidth,
                        "legend.framewidth" = legend.framewidth,
                        "font.size" = font.size,
-                       "rotate_x_axis_labels" = rotate_x_axis_labels)
+                       "rotate_x_axis_labels" = rotate_x_axis_labels,
+                       "sequential_direction" = sequential_direction,
+                       "viridis_direction" = viridis_direction)
   check_type(parameters = numeric_list, required_type = "numeric", test_function = is.numeric)
   # Check character parameters.
   character_list <- list("mode" = mode,
@@ -62,7 +69,9 @@ do_CorrelationPlot <- function(sample,
                          "plot.subtitle" = plot.subtitle,
                          "plot.caption" = plot.caption,
                          "font.type" = font.type,
-                         "diverging.palette" = diverging.palette)
+                         "diverging.palette" = diverging.palette,
+                         "sequential.palette" = sequential.palette,
+                         "viridis.palette" = viridis.palette)
   check_type(parameters = character_list, required_type = "character", test_function = is.character)
 
   check_colors(na.value)
@@ -74,14 +83,19 @@ do_CorrelationPlot <- function(sample,
   check_parameters(parameter = legend.type, parameter_name = "legend.type")
   check_parameters(parameter = number.breaks, parameter_name = "number.breaks")
   check_parameters(parameter = diverging.palette, parameter_name = "diverging.palette")
+  check_parameters(parameter = sequential.palette, parameter_name = "sequential.palette")
 
   `%>%` <- magrittr::`%>%`
 
-  out <- check_and_set_assay(sample = sample, assay = assay)
-  sample <- out[["sample"]]
-  assay <- out[["assay"]]
+  
 
   if (mode == "hvg"){
+    # Check if the sample provided is a Seurat object.
+    check_Seurat(sample = sample)
+    out <- check_and_set_assay(sample = sample, assay = assay)
+    sample <- out[["sample"]]
+    assay <- out[["assay"]]
+    
     if (is.null(group.by)){
       assertthat::assert_that(!("Groups" %in% colnames(sample@meta.data)),
                               msg = paste0(crayon_body("Please, make sure you provide a value for "),
@@ -247,6 +261,148 @@ do_CorrelationPlot <- function(sample,
                         legend.background = ggplot2::element_rect(fill = "white", color = "white"))
 
 
+  } else if (mode == "jaccard"){
+  
+    # Compute jaccard indext.
+    jaccard <- function(set_1, set_2) {
+      # Compute intersection.
+      intersection = length(dplyr::intersect(set_1, set_2))
+      # Compute the union.
+      union = length(set_1) + length(set_2) - intersection
+      # Jaccard index is just the number of shared genes divided by the number of non-shared genes.
+      jaccard_index <- intersection / union
+      return (jaccard_index)
+    }
+    
+    jaccard_scores <- list()
+    for(listname_store in names(input_gene_list)){
+      vector_scores <- c()
+      for(listname in names(input_gene_list)){
+        scores <- jaccard(set_1 = input_gene_list[[listname_store]], set_2 = input_gene_list[[listname]])
+        names(scores) <- listname
+        vector_scores <- c(vector_scores, round(scores, 2))
+      }
+      jaccard_scores[[listname_store]] <- vector_scores
+    }
+    
+    jaccard_matrix <- as.matrix(as.data.frame(jaccard_scores))
+    order <- rownames(jaccard_matrix)[stats::hclust(stats::dist(jaccard_matrix, method = "euclidean"), method = "ward.D")$order]
+    jaccard_matrix <- jaccard_matrix[order, order]
+    jaccard_matrix[jaccard_matrix == 1] <- NA
+    
+    data <- jaccard_matrix %>% 
+            as.data.frame() %>% 
+            tibble::rownames_to_column(var = "x") %>% 
+            tidyr::pivot_longer(cols = -dplyr::all_of(c("x")),
+                                names_to = "y",
+                                values_to = "score") %>% 
+            dplyr::mutate("x" = factor(.data$x, levels = order),
+                          "y" = factor(.data$y, levels = rev(order)))
+    p <- data %>% 
+         ggplot2::ggplot(mapping = ggplot2::aes(x = .data$x,
+                                                y = .data$y,
+                                                fill = .data$score)) +
+         ggplot2::geom_tile(color = "white", linewidth = 0.5, na.rm = TRUE) +
+         ggplot2::scale_y_discrete(expand = c(0, 0)) +
+         ggplot2::scale_x_discrete(expand = c(0, 0),
+                                   position = "top") + 
+         ggplot2::coord_equal() + 
+         ggplot2::guides(y.sec = guide_axis_label_trans(~paste0(levels(.data$y))),
+                         x.sec = guide_axis_label_trans(~paste0(levels(.data$x)))) 
+    limits <- c(min(data$score, na.rm = TRUE),
+                max(data$score, na.rm = TRUE))
+    
+    scale.setup <- compute_scales(sample = NULL,
+                                  feature = NULL,
+                                  assay = NULL,
+                                  reduction = NULL,
+                                  slot = NULL,
+                                  number.breaks = 5,
+                                  min.cutoff = NA,
+                                  max.cutoff = NA,
+                                  flavor = "Seurat",
+                                  enforce_symmetry = FALSE,
+                                  from_data = TRUE,
+                                  limits.use = limits)
+    
+    axis.parameters <- handle_axis(flip = F,
+                                   group.by = "A",
+                                   group = "A",
+                                   counter = 1,
+                                   rotate_x_axis_labels = 45)
+    
+    if (isTRUE(use_viridis)){
+      p <- p + 
+           ggplot2::scale_fill_viridis_c(na.value = na.value,
+                                         option = viridis.palette,
+                                         direction = viridis_direction,
+                                         breaks = scale.setup$breaks,
+                                         labels = scale.setup$labels,
+                                         limits = scale.setup$limits,
+                                         name = "Jaccard score")
+    } else {
+      p <- p + 
+           ggplot2::scale_fill_gradientn(colors = if (sequential_direction == 1){RColorBrewer::brewer.pal(n = 9, name = sequential.palette)[2:9]} else if (sequential_direction == -1){rev(RColorBrewer::brewer.pal(n = 9, name = sequential.palette)[2:9])},
+                                         na.value = na.value,
+                                         name = "Jaccard score",
+                                         breaks = scale.setup$breaks,
+                                         labels = scale.setup$labels,
+                                         limits = scale.setup$limits)
+    }
+   
+    p <- modify_continuous_legend(p = p,
+                                  legend.title = "Jaccard score",
+                                  legend.aes = "fill",
+                                  legend.type = legend.type,
+                                  legend.position = legend.position,
+                                  legend.length = legend.length,
+                                  legend.width = legend.width,
+                                  legend.framecolor = legend.framecolor,
+                                  legend.tickcolor = legend.tickcolor,
+                                  legend.framewidth = 0.5,
+                                  legend.tickwidth = 0.5)
+    
+    p <- p +
+         ggplot2::xlab("") +
+         ggplot2::ylab("") +
+         ggplot2::theme_minimal(base_size = 14) +
+         ggplot2::theme(axis.ticks.x.bottom = axis.parameters$axis.ticks.x.bottom,
+                        axis.ticks.x.top = axis.parameters$axis.ticks.x.top,
+                        axis.ticks.y.left = axis.parameters$axis.ticks.y.left,
+                        axis.ticks.y.right = axis.parameters$axis.ticks.y.right,
+                        axis.text.y.left = axis.parameters$axis.text.y.left,
+                        axis.text.y.right = axis.parameters$axis.text.y.right,
+                        axis.text.x.top = axis.parameters$axis.text.x.top,
+                        axis.text.x.bottom = axis.parameters$axis.text.x.bottom,
+                        axis.title.x.bottom = axis.parameters$axis.title.x.bottom,
+                        axis.title.x.top = axis.parameters$axis.title.x.top,
+                        axis.title.y.right = axis.parameters$axis.title.y.right,
+                        axis.title.y.left = axis.parameters$axis.title.y.left,
+                        strip.background = axis.parameters$strip.background,
+                        strip.clip = axis.parameters$strip.clip,
+                        strip.text = axis.parameters$strip.text,
+                        legend.position = "bottom",
+                        axis.line = ggplot2::element_blank(),
+                        plot.title = ggplot2::element_text(face = "bold", hjust = 0),
+                        plot.subtitle = ggplot2::element_text(hjust = 0),
+                        plot.caption = ggplot2::element_text(hjust = 1),
+                        plot.title.position = "plot",
+                        panel.grid = ggplot2::element_blank(),
+                        panel.grid.minor.y = ggplot2::element_line(color = "white", linewidth = 1),
+                        text = ggplot2::element_text(family = "sans"),
+                        plot.caption.position = "plot",
+                        legend.text = ggplot2::element_text(face = "bold"),
+                        legend.title = ggplot2::element_text(face = "bold"),
+                        legend.justification = "center",
+                        plot.margin = ggplot2::margin(t = 0, 
+                                                      r = 0, 
+                                                      b = 0, 
+                                                      l = 0),
+                        panel.border = ggplot2::element_rect(fill = NA, color = "black", linewidth = 1),
+                        panel.grid.major = ggplot2::element_blank(),
+                        plot.background = ggplot2::element_rect(fill = "white", color = "white"),
+                        panel.background = ggplot2::element_rect(fill = "white", color = "white"),
+                        legend.background = ggplot2::element_rect(fill = "white", color = "white"))
   }
   return(p)
 }